requirements.txt
adafruit-blinka==8.14.0 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-busdevice==5.2.3 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-debouncer==2.0.5 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-framebuf==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-requests==1.12.12 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-rfm9x==2.2.13 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-ssd1306==2.12.12 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-ticks==1.0.9 ; python_version >= "3.9" and python_version < "4.0"
adafruit-circuitpython-typing==1.8.3 ; python_version >= "3.9" and python_version < "4.0"
adafruit-platformdetect==3.40.2 ; python_version >= "3.9" and python_version < "4.0"
adafruit-pureio==1.1.9 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
charset-normalizer==3.0.1 ; python_version >= "3.9" and python_version < "4"
dirsync==2.2.5 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
numpy==1.24.2 ; python_version >= "3.9" and python_version < "4.0"
paho-mqtt==1.6.1 ; python_version >= "3.9" and python_version < "4.0"
pyaudio==0.2.13 ; python_version >= "3.9" and python_version < "4.0"
pydub==0.25.1 ; python_version >= "3.9" and python_version < "4.0"
pyftdi==0.54.0 ; python_version >= "3.9" and python_version < "4.0"
pyserial==3.5 ; python_version >= "3.9" and python_version < "4.0"
python-shout==0.2.7 ; python_version >= "3.9" and python_version < "4.0"
pyusb==1.2.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4"
